Module wasmer_runtime::cache

The cache module provides the common data structures used by compiler backends to allow serializing compiled wasm code to a binary format. The binary format can be persisted, and loaded to allow skipping compilation and fast startup.

Re-exports

pub use super::Backend;

Structs

Artifact
Artifact are produced by caching, are serialized/deserialized to binaries, and contain module info, backend metadata, and compiled code.

FileSystemCache
Representation of a directory that contains compiled wasm artifacts.

WasmHash
The hash of a wasm module.

Traits

Cache
A generic cache for storing and loading compiled wasm modules.