Module wasmer_runtime::memory

The memory module contains the implementation data structures and helper functions used to manipulate and access wasm memory.

Structs

Atomically
Atomically.

Memory
A shared or unshared wasm linear memory.

MemoryView
A view into a memory.