Trait wasmer_runtime::LikeNamespace

pub trait LikeNamespace {
    fn get_export(&self, name: &str) -> Option<Export>;

    fn get_exports(&self) -> Vec<(String, Export)>;

    fn maybe_insert(&mut self, name: &str, export: Export) -> Option<()>;
}

This trait represents objects that act as a namespace for imports. For example, an Instance or ImportObject could be considered namespaces that could provide imports to an instance.

Required methods

fn get_export(&self, name: &str) -> Option<Export>

Gets an export by name.

fn get_exports(&self) -> Vec<(String, Export)>

Gets all exports in the namespace.

fn maybe_insert(&mut self, name: &str, export: Export) -> Option<()>

Maybe insert an Export by name into the namespace.

Loading content...