Module wasmer_runtime::types

Various types.

Structs

FuncIndex
Typed Index

FuncSig
The signature of a function that is either implemented in a wasm module or exposed to wasm by the host.

GlobalDescriptor
Describes the mutability and type of a Global

GlobalIndex
Typed Index

GlobalInit
A wasm global.

ImportedFuncIndex
Typed Index

ImportedGlobalIndex
Typed Index

ImportedMemoryIndex
Typed Index

ImportedTableIndex
Typed Index

LocalFuncIndex
Typed Index

LocalGlobalIndex
Typed Index

LocalMemoryIndex
Typed Index

LocalTableIndex
Typed Index

MemoryDescriptor
A wasm memory descriptor.

MemoryIndex
Typed Index

SigIndex
Index for signature.

TableDescriptor
Describes the properties of a table including the element types, minimum and optional maximum, number of elements in the table.

TableIndex
Typed Index

Enums

ElementType
Kinds of element types.

Initializer
A const value initializer. Over time, this will be able to represent more and more complex expressions.

LocalOrImport
Kind of local or import type.

Type
Represents a WebAssembly type.

Value
Represents a WebAssembly value.

Traits

LocalImport
Trait that represents Local or Import.

NativeWasmType
Represents a native wasm type.

ValueType
Trait for a Value type.

WasmExternType
A trait to represent a wasm extern type.