Enum wasmer_runtime::types::LocalOrImport

pub enum LocalOrImport<T> where
    T: LocalImport,  {
    Local(<T as LocalImport>::Local),
    Import(<T as LocalImport>::Import),
}

Kind of local or import type.

Variants

Local(<T as LocalImport>::Local)

Local.

Import(<T as LocalImport>::Import)

Import.

Methods

impl<T> LocalOrImport<T> where
T: LocalImport,

pub fn local(self) -> Option<<T as LocalImport>::Local>

Returns Some if self is local, None if self is an import.

pub fn import(self) -> Option<<T as LocalImport>::Import>

Returns Some if self is an import, None if self is local.